v1.5 Editor release
1) Added new game mode: editor. In Editor mode you can create new cubes, light cubes, change transparency of each cube (incl. lights) and clear field
2) New game mode available as seperate game (new editor from main menu) or as editor tools during regular maze game

Massive refactoring took place, especially for Field class
