--- maze4d/RectangleGraphics.cpp
@@ -1,12 +1,0 @@
-#include "RectangleGraphics.h"
-
-
-
-RectangleGraphics::RectangleGraphics()
-{
-}
-
-
-RectangleGraphics::~RectangleGraphics()
-{
-}
